+++ apps/web/modules/d/[link]/d-type-view.tsx
@@ -5,7 +5,6 @@ import { useSearchParams } from "next/navigation";
 import { BookerWebWrapper as Booker } from "@calcom/atoms/booker";
 import { getBookerWrapperClasses } from "@calcom/features/bookings/Booker/utils/getBookerWrapperClasses";
 
-import { getUtmTrackingParameters } from "@lib/booking/getUtmTrackingParameters";
 import { type PageProps } from "@lib/d/[link]/[slug]/getServerSideProps";
 
 export default function Type({
@@ -34,7 +33,6 @@ export default function Type({
         duration={duration}
         hashedLink={hashedLink}
         durationConfig={durationConfig}
-        tracking={getUtmTrackingParameters(searchParams)}
       />
     </main>
   );
