+++ apps/web/modules/org/[orgSlug]/instant-meeting/team/[slug]/[type]/instant-meeting-view.tsx
@@ -6,7 +6,6 @@ import { useSearchParams } from "next/navigation";
 import { BookerWebWrapper as Booker } from "@calcom/atoms/booker";
 import { getBookerWrapperClasses } from "@calcom/features/bookings/Booker/utils/getBookerWrapperClasses";
 
-import { getUtmTrackingParameters } from "@lib/booking/getUtmTrackingParameters";
 import type { getServerSideProps } from "@lib/org/[orgSlug]/instant-meeting/team/[slug]/[type]/getServerSideProps";
 import type { inferSSRProps } from "@lib/types/inferSSRProps";
 
@@ -26,7 +25,6 @@ function Type({ slug, user, booking, isEmbed, isBrandingHidden, entity, eventTyp
         isInstantMeeting
         entity={{ ...entity, eventTypeId: eventTypeId }}
         duration={duration}
-        tracking={getUtmTrackingParameters(searchParams)}
       />
     </main>
   );
