+++ apps/web/modules/team/type-view.tsx
@@ -6,7 +6,6 @@ import { useSearchParams } from "next/navigation";
 import { BookerWebWrapper as Booker } from "@calcom/atoms/booker";
 import { getBookerWrapperClasses } from "@calcom/features/bookings/Booker/utils/getBookerWrapperClasses";
 
-import { getUtmTrackingParameters } from "@lib/booking/getUtmTrackingParameters";
 import type { getServerSideProps } from "@lib/team/[slug]/[type]/getServerSideProps";
 import type { inferSSRProps } from "@lib/types/inferSSRProps";
 
@@ -60,7 +59,6 @@ function Type({
         teamMemberEmail={teamMemberEmail}
         crmOwnerRecordType={crmOwnerRecordType}
         crmAppSlug={crmAppSlug}
-        tracking={getUtmTrackingParameters(searchParams)}
       />
     </main>
   );
